requirements.txt
matplotlib==3.8.0
numpy==1.24.1
pillow==10.2.0
torchinfo==1.8.0
torchmetrics==1.5.2
tqdm==4.67.1
wandb==0.19.1